+++ R/knitr.R
@@ -37,7 +37,6 @@ knit_print.huxtable <- function (x, options, ...) {
 
   res <- do.call(call_name, list(x))
 
-  in_quarto <- ! is.null(knitr::opts_knit$get("quarto.version"))
   res <- switch(of,
             latex = {
               latex_deps <- report_latex_dependencies(quiet = TRUE)